brand-grid refactoring
