--- src/components/Cart.tsx
@@ -1,3 +1,0 @@
-export const Cart = () => {
-  return <h1>Cart</h1>;
-};
